Make it possible to return a C++ `ShardedDeviceArray`.
This _will_ be a *breaking* change, as pxla.ShardedDeviceArray constructor won't be valid anymore, when:
- for the next Jax release
- with a new `jaxlib` with _xla_extension_version=29
- on the condition _USE_EXPERIMENTAL_CPP_SDA is switch to True.

I am already adding the impact to the user in the CHANGELOG, we can still move it to the next version.

Similarly to JAX.jit, for which we have a C++ `DeviceArray` and a Python `_DeviceArray`, we will introduce 2 objects for ShardedDeviceArray, with the Python object only for JAX extensions not compatible with the C++ object (e.g. Cloud TPU).

- Add `make_sharded_device_array` to be used within JAX and for hackers that need to construct SDA objects.
- Make sure the C++ object is valid by
  (a) extending `DeviceArrayBase` (done in Python), as it brings a bunch of methods and enable `isinstance(x, DeviceArray)`
  (b) Adding the same methods as the Python SDA.

NOTE: mypy has troubled with the " -> pxla.ShardedDeviceArray` function return type annotation, I had to remove 2.
PiperOrigin-RevId: 388190299